+++ src/core/tools/tools.hpp
@@ -4,4 +4,4 @@
 
 // todo rename to timeTools.hpp and timeTools.cpp
 
-unsigned long secondsToNextHourStart(Date now);
+unsigned int secondsToNextHourStart(Date now);
